Semana 14 - Entidades Realcionadas

+++ semana-13/backend/src/controllers/PointController.js
@@ -2,7 +2,9 @@ import Point from "../models/PointModel.js";
 
 export const getPoints = async (req, res) => {
     try {
-        const points = await Point.find().sort({ createdAt: -1 });
+        const points = await Point.find()
+                                    .populate("createdBy", "name email")
+                                    .sort({ createdAt: -1 });
         res.json(points);
     } catch (error) {
         res.status(500).json({ message: "Error al obtener los puntos" });
@@ -25,7 +27,12 @@ export const getPointById = async (req, res) => {
 
 export const createPoint = async (req, res) => {
   try {
-    const { title, description, category, lat, lng } = req.body;
+    const { title, lat, lng } = req.body;
+
+
+    const { _id} = req.user;
+
+    console.log({ _id})
 
     if (!title  || lat === undefined || lng === undefined) {
         return res.status(400).json({
@@ -35,33 +42,45 @@ export const createPoint = async (req, res) => {
 
     const newPoint = new Point({
         title,
-        description,
-        category,
         lat,
-        lng
+        lng,
+        createdBy: _id
     });
 
     await newPoint.save();
 
     res.status(201).json(newPoint);
   } catch (error) {
+        console.error(error);
         res.status(500).json({ message: "Error al crear el punto" });
   }
 };
 
 export const updatePoint = async (req, res) => {
     try {
-        const point = await Point.findByIdAndUpdate(
-            req.params.id,
-            req.body,
-            { new: true }
-        );
-
+        const point = await Point.findById(req.params.id );
         if (!point) {
             return res.status(404).json({ message: "Punto no encontrado" });
         }
+        // Solo el Creador o el admin puede editar
+        const isOwner = point.createdBy.toString() === req.user._id.toString();
+        const idAdmin = req.user.role == "admin";
 
-        res.json(point);
+        if( !isOwner && !idAdmin ){
+            return res.status(403).json({ 
+                message: "No tenés permisos para modificar" 
+            });
+        }
+
+        const updatePoint = await Point.findByIdAndUpdate(
+            req.params.id,
+            req.body,
+            { new: true}
+        )
+
+
+
+        res.json(updatePoint);
     } catch (error) {
         res.status(500).json({ message: "Error al actualizar el punto" });
     }
@@ -69,11 +88,21 @@ export const updatePoint = async (req, res) => {
 
 export const deletePoint = async (req, res) => {
     try {
-        const point = await Point.findByIdAndDelete(req.params.id);
+
+        const point = await Point.findById(req.params.id );
 
         if (!point) {
             return res.status(404).json({ message: "Punto no encontrado" });
         }
+
+        // Solo el Creador o el admin puede Eliminiar
+        if( req.user.role !== "admin"){
+            return res.status(403).json({ 
+                message: "No tenés permisos para eliminar" 
+            });
+        }
+
+        await Point.findByIdAndDelete( req.params.id);
 
         res.json({ message: "Punto eliminado correctamente" });
     } catch (error) {
